refactor: 참여자들이 가지고 있는 카드를 클래스로 분리
