+++ src/main/java/domain/Player.java
@@ -9,6 +9,6 @@ public class Player extends BlackjackParticipant {
     @Override
     boolean isDrawable() {
         int sum = calculateCardSum();
-        return !isBust(sum);
+        return hand.isBust(sum);
     }
 }
